third_party/WebKit/LayoutTests/fast/forms/calendar-picker/datetimelocal-picker-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type="datetime-local" id="datetimelocal1" value="">

<script>
description('Tests if value selection by calendar picker dispatches correct events.');

var eventsCounter = {};
function recordEvent(event) {
    if (eventsCounter[event.type] === undefined)
        eventsCounter[event.type] = 0;
    eventsCounter[event.type]++;
    debug('==> "' + event.type + '" event was dispatched.');
}

var datetimelocal1 = document.getElementById('datetimelocal1');
datetimelocal1.addEventListener('input', recordEvent, false);
datetimelocal1.addEventListener('change', recordEvent, false);

openPicker(datetimelocal1, test1);

function test1() {
    debug('Choosing a new date value from the calendar picker. No events should be dispatched because the hour field and the minutes field are empty.');
    removeCommitDelay();
    eventSender.keyDown('ArrowLeft');
    eventSender.keyDown('Enter');
    shouldBeEqualToString('datetimelocal1.value', '');
    shouldBeUndefined('eventsCounter.input');
    shouldBeUndefined('eventsCounter.change');

    waitUntilClosing(test1AfterClosing);
}

function test1AfterClosing() {
    datetimelocal1.value = "2013-01-21T17:49";
    openPicker(datetimelocal1, test2);
}

function test2() {
    debug('Choosing a new value from the calendar picker. "Input" and "change" events should be dispatched in this order.');
    eventSender.keyDown('ArrowRight');
    removeCommitDelay();
    eventSender.keyDown('Enter');
    shouldBeEqualToString('datetimelocal1.value', '2013-01-22T17:49');
    shouldBe('eventsCounter.input', '1');
    shouldBe('eventsCounter.change', '1');

    finishJSTest();
}
</script>
</body>
</html>
